WIP: many changes to processing routines
Numerous changes to my processing routines on KAZR, XSAPR, CSAPR, and
NEXRAD WSR-88D data. Below is a brief list of the changes involved in
this commit.

1. Adding radar significant detection mask in WSR-88D CMAC files. Since
the NEXRAD data is already processed to some degree (e.g., data not
available at every gate due to clutter filtering etc.), the significant
detection is any gate not missing (as defined by the reflectivity
field). The reason for adding significant detection is that it offers a
signal quality index (SQI) field for the radar data objective analysis
(gridding), i.e., the grid quality index.

2. Some tuning of dealiasing and significant detection parameters for
the ARM radars in order to improve results. This includes (for CSAPR)
the addition of my significant echo boundaries routine.